requirements.txt
pinecone-haystack
haystack-ai
fastapi
uvicorn
python-dotenv
pathlib
pypdf
nltk>=3.9.1
sentence-transformers

-e .